pyproject.toml
[project]
name = "langchain-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "ipykernel>=7.2.0",
    "isort>=7.0.0",
    "langchain>=0.0.27",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.15",
    "langchain-text-splitters>=1.1.0",
    "pip>=26.0.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
